+++ packages/pass-style/test/errors.test.js
@@ -32,6 +32,11 @@ test('style of extended errors', t => {
     const a4 = harden(AggregateError([e2, u3], 'a4', { cause: e1 }));
     t.is(passStyleOf(a4), 'error');
   }
+  if (typeof SuppressedError !== 'undefined') {
+    // Conditional, to accommodate platforms prior to SuppressedError
+    const a4 = harden(SuppressedError(e2, u3, 'a4'));
+    t.is(passStyleOf(a4), 'error');
+  }
 });
 
 test('toPassableError, toThrowable', t => {