+++ packages/ses/NEWS.md
@@ -2,6 +2,8 @@ User-visible changes in `ses`:
 
 # v1.9.0 (2024-10-10)
 
+- TODO explain `SuppressedError` support
+
 - On platforms without
   [`Array.prototype.transfer`](https://github.com/tc39/proposal-resizablearraybuffer)
   but with a global `structuredClone`, the ses-shim's `lockdown` will now